+++ hc/api/models.py
@@ -1048,6 +1048,7 @@ class Channel(models.Model):
             "whatsapp",
             "ntfy",
             "group",
+            "gotify",
         )
 
     def assign_all_checks(self) -> None: