+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - Send "Monthly Phone/SMS/WA Limit Reached" notice to all team members (not just owner)
 - Include notification text in the "Monthly Phone/SMS/WA Limit Reached" notice
 - Add a note about unset ping key in the check details page (#1295)
+- In the "Add Gotify" form, let users specify alert priorities (#990, @BeyondMagic)
+- Add ability to edit existing Gotify integrations
 
 ### Bug Fixes
 - Fix WebhookValidator to handle input that contains square brackets